Add screen/window layout options
This should not change the default behavior (fullscreen) for any display
resolution that is <= 1280x720.

For larger displays, defaults are changed to show a window with a frame.
The position and size of the window is tracked and restored between
executions. This is a better behavior for running OctoDash on a desktop
display.

The default behavior for "--serve" mode is also changed.

Window optons have been added to settings to allow customization of:
    x, y, width, height, fullscreen, frame, and backgroundColor
These customizations are disabled by default and may be changed by the
user in settings->octodash.