Add [!IMG] custom action and refactor IFrame code
MJPEG webcams can be viewed with the <img> tag. The advantage over an
<iframe> is that formatting of the video stream can be controlled
through css. It can be scaled to fit the screen while maintaining proper
aspect ratio. High resolution video results in scrollbars in the
<iframe>.

Also refactor WEB/IMG into it's own routed Component. The new component
has a "Back" button that returns to where it was invoked from. There is
an overlay that displays printer/job status. The notification-center is
accessible from the component.

FYI, sharing the IFrame as part of the custom actions template between
notification-center and control resulted in "showing" the IFrame in a
context that wasn't visible because getElementById returns the first
match it finds and the id being searched occurred twice.